+++ CHANGELOG.md
@@ -34,6 +34,11 @@ Change categories:
 
 - base: Add encode req & decode resp for MultipartReceive
 
+- pdr: Add pldm_file_descriptor_pdr struct
+
+- platform: Add decode_pldm_file_descriptor_pdr() and
+  decode_pldm_file_descriptor_pdr_names()
+
 ### Changed
 
 - dsp: firmware_update: Expand "params" in symbol names